각종 에러 수정

+++ database/migrations/2020_06_17_081237_create_comment_table.php
@@ -14,7 +14,7 @@ class CreateCommentTable extends Migration
     public function up()
     {
         Schema::create('comment', function (Blueprint $table) {
-            $table->BigInteger('comment_num')->unique();
+            $table->bigIncrements('comment_num')->unique();
             $table->date('time');
             $table->string('comments');
             $table->timestamps();
